pyproject.toml
[project]
name = "doc-automation"
version = "0.1.0"
description = "AI Document Processor Microservice"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "streamlit (>=1.45.1,<2.0.0)",
    "azure-ai-formrecognizer (>=3.3.3,<4.0.0)",
    "azure-ai-documentintelligence (>=1.0.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
